pyproject.toml
[project]
name = "prove"
version = "0.2.0"
description = "Minimal voice cloning system using audio reference"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "chatterbox-tts>=0.1.4",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "pydub>=0.25.1",
    "numpy>=1.24.3,<2.0.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.3",
    "black>=24.1.0",
    "ruff>=0.1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
